awful.rules: Allow rules to be added and removed.
Begin to move away from the "write once" rules. This is part of the
project to make `rc.lua` more modular. It will be possible to add rules
from multiple modules. While it could be done with sources, this is a
different use case. The sources have priorities while this is more about
rules within a source.